requirements.txt
streamlit
ultralytics
pillow
gtts
speechrecognition
anthropic
pyaudio